grid_generation_vina-main/requirements.txt
biopython==1.84
numpy==2.1.1
